[PP-2841] Ensure that the pool with the highest license availability … (#2854)
…is displayed in the in the OPDS feed.

## Description
This update ensures that the availability information returned to the
Admin UI as well as the mobile app in the OPDS feed for a work returns
the most available license pool's availability information. Previously
the Work.active_license_pool method was not taking into account the
availability when determining the active_license_pool. As a result, the
availability counts being returned in the feed were inconsistent and in
some case wrong. It it was possible for a work that has available
licenses to appear not to. With the changes in the PR this issue should
be addressed as well as making the active_license_pool result more
deterministic.

## Motivation and Context
https://ebce-lyrasis.atlassian.net/browse/PP-2841
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
Unit tests have been added.  
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.