Fix outdated documentation. (#2869)
## Description
Fixes out of date docstring. This method was altered in a recent PR
review cycle but the docs were not properly updated after I modified the
implementation of the method. This is a follow on PR to correct a
problem that fell through the cracks in previous PR. It was small enough
that it didn't warrant a separate ticket (IMHO).

Here's where the code was originally introduced - you can see the
comments are out of date (because the implementation changed in the
course of the PR)

https://github.com/ThePalaceProject/circulation/pull/2851/files#diff-b57915e62c6c23ce05aa81fcfd0f9e22d759df6b28c18461fa91e0fa2d13d9b7R313-R338
<!--- Describe your changes -->

## Motivation and Context
The docs were out of sync with the code.
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
No code changes to test.

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.