[PP-3190] Respect library identifier field in sirsidynix horizon pro… (#2870)
…vider.

## Description
Before this fix, the Horizon provider was assuming the presence of a
"patronType" in the patron data returned from the sirsi dynix provider
even if the library identifier field in the settings was set to
"barcode". This fix updates the provider to respect the library
identifier field. Additionally in the process of fixing this issue I ran
into another problem which is that at least some Symphony services
provide a "standing" field rather than "approved" to indicate whether a
user is in good standing and thus approved. I added a test to cover this
fix as well as a TODO in case we need to make this configurable. With
the changes, the user must be approved or have a standing of "OK" (case
insensitive) to avoid being blocked.

## Motivation and Context
https://ebce-lyrasis.atlassian.net/browse/PP-3190
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
Unit tests added and manually tested the fix locally.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.