[PP-3189] prefix only authetication provider (#2871)
## Description
The first thing I did was to add an option to the
BasicAuthenticationProvider that allows users to restrict based on one a
list of prefixes. Then I added a new authentication provider,
MinimalAuthenticationProvider, which simply extends the
BasicAuthenticationProvider and implemented the two abstract methods. As
a descendent of BasicAuthenticationProvider, it provides all the local
validation options (including not presenting a password, restricting to
a list of prefixes, etc) .


## Motivation and Context

[PP-3189](https://ebce-lyrasis.atlassian.net/browse/PP-3189)

<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
I added unit test coverage and tested locally. I verified that it is
possible to restrict successful authentication to a list of user defined
prefixes.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.


[PP-3189]:
https://ebce-lyrasis.atlassian.net/browse/PP-3189?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Tim DiLauro <tdilauro@users.noreply.github.com>