[PP-3347] Reduce the default page size on overdrive api fetch  to 25 … (#2912)
…from 100.

This update reduces the pages size and therefore the rate of circulation
update task generation by 75%. The hope is that this adjustment will
reduce the rate at which new bibliographic and circulation apply tasks
are created, thereby preventing the rapid consumption of available
memory in the redis instances as it will hopefully give the apply queue
a chance to get consumed.
## Motivation and Context
https://ebce-lyrasis.atlassian.net/browse/PP-3347
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
I plan to test this change by cutting a bugfix release and slowly
rolling it out to the production instances, cm by cm on a specific shard
while monitoring the redis memory footrpint and overdrive error rate.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.