[PP-3376] Increase nginx large client buffers to support long quicksi… (#2919)
…ght embedded url generator endpoint query strings.

Quicksight Dashboard was failing on Illinois only due to a 413 error
coming back from our webservers when trying to get an embedded
quicksight url. Illinois unlike other CMs has many many more libraries.
As a result the URL we build to generate a quicksight embedded url is
much longer.

This update increases the limit from 8k to 24k which should be plenty
since the URL that generated the error was about 12 k.

https://ebce-lyrasis.atlassian.net/browse/PP-3376