[PP-3215] Revert remove old overdrive scripts (#2923)
This reverts commit 598fcf4709c00a3e90e3197a8bd09796fb50d981.

## Description

This commit turned on the celery based overdrive import. There are still
issues that need to be resolved before this can go live. Once those
issues are resolved we can revert this reversion.

## Motivation and Context

See: #2889

## How Has This Been Tested?

This reversion tested in production since the current deployments to
prod do not have this commit.

## Checklist

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.