[PP-3377] Prevent/abort import if flagged for deletion. (#2918)
## Description
If an Overdrive collection is flagged for deletion while it is being
imported the import should cease asap.

This PR adds a check for the collection's marked_for_deletion_flag
before importing.

If this update corrects the problem in production, I will create follow
on ticket to address the other import tasks (such as OPDS2).


## Motivation and Context
https://ebce-lyrasis.atlassian.net/browse/PP-3377
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.