[PP-3376] Makes library_uuids param optional for generating quicksigh… (#2936)
…t urls. If not present, all allowable libraries are used to generate
embed url.

As a matter of fact, this capability was in place prior to this change,
the validation constraints prevented us from making use of it. This PR
changes the validation rules make the library_uuids parameter optional.
When not present it is assumed that all allowable libraries will be
included in the quicksight dashboard.
## Motivation and Context
This change is necessary to prevent the circulation admin from sending
excessively long urls which blow past cloudfront and nginx limits to the
CM.

https://ebce-lyrasis.atlassian.net/browse/PP-3376

## How Has This Been Tested?
New unit test coverage.


## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.